Python Project
This notebook processes and cleans data related to ultramarathon races in the U.S. during 2020, focusing on 50-mile and 50-kilometer events. It filters the dataset to include only relevant races, removes unnecessary columns, handles missing values, and calculates new information like athlete age. The data types are corrected, and column names are renamed for clarity. It is a clean and organized dataset ready for further analysis.